move regression functions to separate module
